[python] Overhaul iree.build console output and error handling.
* Status updates look more like what would be expected from something like ninja or bazel:
  * Short update one liners wink past
  * If some actions are taking a long time, a multi-line display is shown with the worst lagging ones
* Per action errors and dependence errors are now tracked in the build graph and reported correctly (vs triggering the last chance exception handler).
* The last chance exception handler for true program errors now prints its exception verbosely prior to blocking waiting for the executor to shutdown (this was causing exception swallowing).
